+++ src/components/ui/alert-dialog.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog"
 
 import { cn } from "@/lib/utils"
-import { buttonVariants } from "@/components/ui/button"
+import { buttonVariants } from "@/components/ui/button-helpers"
 
 const AlertDialog = AlertDialogPrimitive.Root
